Use bundled matchit.
Vim's bundled matchit lives in the macros/ directory, whence it's not
auto-loaded.  We have to load it ourselves.

As and when a new version becomes available via vim.org, it'll be
installed into the ~/.vim/plugin/ directory.  Files there are
auto-loaded so we'll be able to remove the reference to it from vimrc.